requirements.txt
asgiref==3.10.0
celery==5.3.4
Django==4.2.25
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
dotenv==0.9.9
openpyxl==3.1.5
pandas==2.2.3
pillow==11.3.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
python-dateutil==2.9.0
python-dotenv==1.2.1
pytz==2024.1
redis==5.0.1
sqlparse==0.5.3
typing_extensions==4.15.0
